pyproject.toml
[project]
name = "disasterchat"
version = "0.1.0"
description = ""
authors = [
    {name = "wzx",email = "772618452wang@gmail.com"}
]
readme = "README.md"
requires-python = "<=3.13,>=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "ollama (>=0.1.5,<0.2.0)",
    "fastapi-cors (>=0.0.6,<0.0.7)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

